[STYLE] clean code about dataframe utils (#816)
* [STYLE] clean code

* use to_json instead of to_dict
