+++ bentoml/adapters/fastai_image_input.py
@@ -16,13 +16,20 @@ import os
 import argparse
 import base64
 from io import BytesIO
+import json
 
 from werkzeug.utils import secure_filename
 from flask import Response
 import numpy as np
 
+try:
+    import pandas as pd
+except ImportError:
+    pd = None
+
+from bentoml.utils.dataframe_util import PANDAS_DATAFRAME_TO_JSON_ORIENT_OPTIONS
 from bentoml.exceptions import BadInput, MissingDependencyException
-from bentoml.adapters.base_input import BaseInputAdapter, api_func_result_to_json
+from bentoml.adapters.base_input import BaseInputAdapter
 from bentoml.adapters.image_input import (
     verify_image_format_or_raise,
     get_default_accept_image_formats,
@@ -49,6 +56,38 @@ def _import_imageio_imread():
         )
 
     return imread
+
+
+class NumpyJsonEncoder(json.JSONEncoder):
+    """ Special json encoder for numpy types """
+
+    def default(self, o):  # pylint: disable=method-hidden
+        if isinstance(o, np.generic):
+            return o.item()
+
+        if isinstance(o, np.ndarray):
+            return o.tolist()
+
+        return json.JSONEncoder.default(self, o)
+
+
+def api_func_result_to_json(result, pandas_dataframe_orient="records"):
+
+    assert (
+        pandas_dataframe_orient in PANDAS_DATAFRAME_TO_JSON_ORIENT_OPTIONS
+    ), f"unkown pandas dataframe orient '{pandas_dataframe_orient}'"
+
+    if pd and isinstance(result, pd.DataFrame):
+        return result.to_json(orient=pandas_dataframe_orient)
+
+    if pd and isinstance(result, pd.Series):
+        return pd.DataFrame(result).to_json(orient=pandas_dataframe_orient)
+
+    try:
+        return json.dumps(result, cls=NumpyJsonEncoder)
+    except (TypeError, OverflowError):
+        # when result is not JSON serializable
+        return json.dumps({"result": str(result)})
 
 
 class FastaiImageInput(BaseInputAdapter):
